View Slider, orientation resize

+++ Source/Clock.h
@@ -39,11 +39,11 @@ public:
     
 protected:
     bool isPlaying = false;
-    float bpm = 120;
+    double bpm = 120.0;
     int cursor = -1;
     int numCells = 16;
-    float sixteenthTimeMs = 60000 / bpm / 4;
-    float lastClockStepTimeMs = 0;
+    double sixteenthTimeMs = 60000.0 / bpm / 4.0;
+    double lastClockStepTimeMs = 0;
     OwnedArray<ClockListener> listeners;
     OwnedArray<Parameter>* parameters;
     Mixer& mixer;
